backend/requirements.txt
asgiref==3.8.1
Django==4.2.23
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
djangorestframework==3.16.0
drf-spectacular==0.27.1
exceptiongroup==1.3.0
iniconfig==2.1.0
mysqlclient==2.2.7
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-django==4.11.1
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.14.0
